Control.Arrow.ArrowNF

Description

Arrows for evaluation of normal form results

Documentation

strictA :: (Arrow a, NFData b) => a b b

complete evaluation of an arrow result using DeepSeq

this is sometimes useful for preventing space leaks, especially after reading and validation of a document, all DTD stuff is not longer in use and can be recycled by the GC.

class Arrow a => ArrowNF a where

Minimal complete definition

Nothing

Methods

rnfA :: NFData c => a b c -> a b c

class (Arrow a, ArrowList a) => ArrowWNF a where

partial evaluation of an arrow result using FlatSeq

There are two arrows with force the partial evaluation. By convention the 2. should be less lazy than the 1.

These arrows are sometimes useful for preventing space leaks, especially when parsing complex data structures. In many cases the evaluated AST is more space efficient than the unevaluaded with a lot of closures.

Minimal complete definition

Nothing

Methods

rwnfA :: WNFData c => a b c -> a b c

rwnf2A :: WNFData c => a b c -> a b c