Control.Arrow.StateListArrow

Description

Implementation of list arrows with a state

Documentation

newtype SLA s a b

list arrow combined with a state

Constructors

SLA

Fields

runSLA :: s -> a -> (s, [b])

fromSLA :: ArrowList a => s -> SLA s b c -> a b c

conversion of state list arrows into arbitray other list arrows.

allows running a state list arrow within another arrow:

example:

... >>> fromSLA 0 (... setState ... getState ... ) >>> ...

runs a state arrow with initial state 0 (e..g. an Int) within another arrow sequence