Control.FlatSeq

Description

Force evaluation like deepseq in Control.DeepSeq, but control the depth of evaluation. flatseq may evaluate more than seq but less than deepseq

Documentation

($!!) :: WNFData a => (a -> b) -> a -> b infixr 0

flatseq :: WNFData a => a -> b -> b

rlnf :: (a -> ()) -> [a] -> ()

class WNFData a where

A class of types that can be partially evaluated, but evaluation can be propagated deeper than WHNF

Minimal complete definition

Nothing

Methods

rwnf :: a -> ()

Default for rwnf is reduction to WHNF

rwnf2 :: a -> ()

Default for rwnf2 is rwnf