Data.AssocList

Description

Simple key value assocciation list implemented as unordered list of pairs

Documentation

type AssocList k v = [(k, v)]

lookupDef :: Eq k => v -> k -> AssocList k v -> v

lookup with default value

lookup1 :: Eq k => k -> AssocList k [e] -> [e]

lookup with empty list (empty string) as default value

hasEntry :: Eq k => k -> AssocList k v -> Bool

test for existence of a key

addEntry :: Eq k => k -> v -> AssocList k v -> AssocList k v

add an entry, remove an existing entry before adding the new one at the top of the list, addEntry is strict

addEntries :: Eq k => AssocList k v -> AssocList k v -> AssocList k v

add a whole list of entries with addEntry

delEntry :: Eq k => k -> AssocList k v -> AssocList k v

delete an entry, delEntry is strict

delEntries :: Eq k => [k] -> AssocList k v -> AssocList k v

delete a list of entries with delEntry