Data.Function.Selector

Documentation

data Selector s a

A Selector is a pair of an access function and a modifying function for reading and updating parts of a composite type

Constructors

S

Fields

getS :: s -> a

setS :: a -> s -> s

chgS :: Selector s a -> (a -> a) -> s -> s

chgM :: Monad m => Selector s a -> (a -> m a) -> s -> m s

mkSelector :: (s -> a) -> (a -> s -> s) -> Selector s a

Alias for constructor S

idS :: Selector s s

(.&&&.) :: Selector s a -> Selector s b -> Selector s (a, b) infixr 3

class Comp1 s a | s -> a where

Selectors for pairs and 3-tuples: comp1, comp2, comp3, this can be extended to n-tuples

Methods

comp1 :: Selector s a

class Comp2 s a | s -> a where

Methods

comp2 :: Selector s a

class Comp3 s a | s -> a where

Methods

comp3 :: Selector s a