Text.XML.HXT.Arrow.Binary

Description

De-/Serialisation arrows for XmlTrees and other arbitrary values with a Binary instance

Documentation

readBinaryValue :: Binary a => String -> IOStateArrow s b a

writeBinaryValue :: Binary a => String -> IOStateArrow s a ()

Serialize a value, optionally compress it, and write it to a file. In case of an error, the error message is issued and the arrow fails