Text.XML.HXT.Arrow.DocumentOutput

Description

State arrows for document output

Documentation

putXmlDocument :: Bool -> String -> IOStateArrow s XmlTree XmlTree

Write the contents of a document tree into an output stream (file or stdout).

If textMode is set, writing is done with Haskell string output, else (default) writing is done with lazy ByteString output

putXmlTree :: String -> IOStateArrow s XmlTree XmlTree

write the tree representation of a document to a file

putXmlSource :: String -> IOStateArrow s XmlTree XmlTree

write a document with indentaion and line numers

encodeDocument :: Bool -> Bool -> String -> IOStateArrow s XmlTree XmlTree

encodeDocument' :: ArrowXml a => Bool -> Bool -> String -> a XmlTree XmlTree