Text.XML.HXT.Arrow.XmlOptions

Description

system configuration and common options options

Documentation

inputOptions :: [OptDescr SysConfig]

commonly useful options for XML input

can be used for option definition with haskell getopt

defines options: a_trace, a_proxy, a_use_curl, a_do_not_use_curl, a_options_curl, a_encoding, a_issue_errors, a_do_not_issue_errors, a_parse_html, a_parse_by_mimetype, a_issue_warnings, a_do_not_issue_warnings, a_parse_xml, a_validate, a_do_not_validate, a_canonicalize, a_do_not_canonicalize,

outputOptions :: [OptDescr SysConfig]

commonly useful options for XML output

defines options: a_indent, a_output_encoding, a_output_html and others

generalOptions :: [OptDescr SysConfig]

commonly useful options

defines options: a_verbose, a_help

versionOptions :: [OptDescr SysConfig]

defines a_version option

showOptions :: [OptDescr SysConfig]

debug output options

a_accept_mimetypes :: String

a_add_default_dtd :: String

a_canonicalize :: String

a_check_namespaces :: String

a_collect_errors :: String

a_default_baseuri :: String

a_do_not_canonicalize :: String

a_do_not_check_namespaces :: String

a_do_not_issue_errors :: String

a_do_not_issue_warnings :: String

a_do_not_preserve_comment :: String

a_do_not_remove_whitespace :: String

a_do_not_subst_dtd_entities :: String

a_do_not_subst_html_entities :: String

a_do_not_validate :: String

a_error :: String

a_error_log :: String

a_help :: String

a_if_modified_since :: String

a_if_unmodified_since :: String

a_ignore_encoding_errors :: String

a_ignore_none_xml_contents :: String

a_indent :: String

a_issue_errors :: String

a_issue_warnings :: String

a_mime_types :: String

a_no_empty_elements :: String

a_no_empty_elem_for :: String

a_no_redirect :: String

a_no_xml_pi :: String

a_output_file :: String

a_output_xml :: String

a_output_html :: String

a_output_xhtml :: String

a_output_plain :: String

a_parse_by_mimetype :: String

a_parse_html :: String

a_parse_xml :: String

a_preserve_comment :: String

a_proxy :: String

a_redirect :: String

a_remove_whitespace :: String

a_show_haskell :: String

a_show_tree :: String

a_strict_input :: String

a_subst_dtd_entities :: String

a_subst_html_entities :: String

a_text_mode :: String

a_trace :: String

a_validate :: String

a_verbose :: String

selectOptions :: [String] -> [OptDescr a] -> [OptDescr a]

select options from a predefined list of option descriptions

removeOptions :: [String] -> [OptDescr a] -> [OptDescr a]