Text.XML.HXT.Arrow.XmlState.RunIOStateArrow

Description

run an io state arrow

Documentation

runX :: IOSArrow XmlTree c -> IO [c]

apply an IOSArrow to an empty root node with initialState () as initial state

the main entry point for running a state arrow with IO

when running runX f an empty XML root node is applied to f. usually f will start with a constant arrow (ignoring the input), e.g. a readDocument arrow.

for usage see examples with writeDocument

if input has to be feed into the arrow use runIOSLA like in runIOSLA f emptyX inputDoc

runXIOState :: XIOState s -> IOStateArrow s XmlTree c -> IO [c]

initialState :: us -> XIOState us

the default global state, used as initial state when running an IOSArrow with runIOSLA or runX

initialSysState :: XIOSysState

initialSysWriter :: XIOSysWriter

initialSysEnv :: XIOSysEnv

initialInputConfig :: XIOInputConfig

initialParseConfig :: XIOParseConfig

initialOutputConfig :: XIOOutputConfig

initialRelaxConfig :: XIORelaxConfig

initialXmlSchemaConfig :: XIOXmlSchemaConfig

initialCacheConfig :: XIOCacheConfig

dummyHTTPHandler :: IOSArrow XmlTree XmlTree

dummyTagSoupParser :: IOSArrow b b

dummyExpatParser :: IOSArrow b b

dummyRelaxValidator :: IOSArrow b b

dummyXmlSchemaValidator :: IOSArrow b b

dummyCacheRead :: String -> IOSArrow b b

getConfigAttr :: String -> SysConfigList -> String

theSysConfigComp :: Selector XIOSysState a -> Selector SysConfig a