Text.XML.HXT.DOM.ShowXml

Description

XML tree conversion to external string representation

Documentation

xshow :: XmlTrees -> String

convert a list of trees into a string

see also : xmlTreesToText for filter version, xread for the inverse operation

xshowBlob :: XmlTrees -> Blob

convert an XML tree into a binary large object (a bytestring)

xshow' :: (Char -> StringFct) -> (Char -> StringFct) -> (Char -> StringFct) -> XmlTrees -> Blob

convert a list of trees into a blob.

Apply a quoting function for XML quoting of content, a 2. quoting funtion for attribute values and an encoding function after tree conversion

xshow'' :: (Char -> StringFct) -> (Char -> StringFct) -> XmlTrees -> String