Text.XML.HXT.DOM.TypeDefs

Description

The core data types of the HXT DOM.

Documentation

module Data.AssocList

type ChildNodes = [(Int, XmlNodeSet)]

data XmlNodeSet

data type for representing a set of nodes as a tree structure

this structure is e.g. used to repesent the result of an XPath query such that the selected nodes can be processed or selected later in processing a document tree

Constructors

XNS

Fields

thisNode :: Bool

is this node part of the set ?

attrNodes :: [QName]

the set of attribute nodes

childNodes :: ChildNodes

the set of child nodes, a list of pairs of index and node set

type Attributes = AssocList String String

Attribute list

used for storing option lists and features of DTD parts

type Blob = ByteString

Binary large object implemented as a lazy bytestring

data DTDElem

Represents a DTD element

Constructors

DOCTYPE
attr: name, system, public, XDTD elems as children

ELEMENT
attr: name, kind

name: element name

kind: "EMPTY" | "ANY" | "#PCDATA" | children | mixed

CONTENT
element content

attr: kind, modifier

modifier: "" | "?" | "*" | "+"

kind: seq | choice

ATTLIST
attributes: name - name of element

value - name of attribute

type: "CDATA" | "ID" | "IDREF" | "IDREFS" | "ENTITY" | "ENTITIES" |

"NMTOKEN" | "NMTOKENS" |"NOTATION" | "ENUMTYPE"

kind: "IMPLIED" | "DEFAULT"

ENTITY
for entity declarations

PENTITY
for parameter entity declarations

NOTATION
for notations

CONDSECT
for INCLUDEs, IGNOREs and peRefs: attr: type

type = INCLUDE, IGNORE or %...;

NAME
attr: name

for lists of names in notation types or nmtokens in enumeration types

PEREF
for Parameter Entity References in DTDs

data XNode

Represents elements

Constructors

XText String
ordinary text (leaf)

XBlob Blob
text represented more space efficient as bytestring (leaf)

XCharRef Int
character reference (leaf)

XEntityRef String
entity reference (leaf)

XCmt String
comment (leaf)

XCdata String
CDATA section (leaf)

XPi QName XmlTrees
Processing Instr with qualified name (leaf) with list of attributes. If tag name is xml, attributes are "version", "encoding", "standalone", else attribute list is empty, content is a text child node

XTag QName XmlTrees
tag with qualified name and list of attributes (inner node or leaf)

XDTD DTDElem Attributes
DTD element with assoc list for dtd element features

XAttr QName
attribute with qualified name, the attribute value is stored in children

XError Int String
error message with level and text

type XmlNavTrees = [NTZipper XNode]

List of navigatable rose trees with XML nodes

type XmlNavTree = NTZipper XNode

Navigatable rose tree with XML nodes

type XmlTrees = NTrees XNode

List of rose trees with XML nodes

type XmlTree = NTree XNode

Rose tree with XML nodes (XNode)

rwnfAttributes :: Attributes -> ()

Evaluate an assoc list of strings

blobToString :: Blob -> String

stringToBlob :: String -> Blob

c_ok :: Int

no error, everything is ok

c_warn :: Int

Error level for XError, type warning

c_err :: Int

Error level for XError, type error

c_fatal :: Int

Error level for XError, type fatal error

module Text.XML.HXT.DOM.QualifiedName