Text.XML.HXT.DTDValidation.DocTransformation

Description

This module provides functions for transforming XML documents represented as XmlTree with respect to its DTD.

Transforming an XML document with respect to its DTD means:

add all attributes with default values

normalize all attribute values

sort all attributes in lexical order

Note: Transformation should be started after validation.

Before the document is validated, a lookup-table is build on the basis of the DTD which maps element names to their transformation functions. After this initialization phase the whole document is traversed in preorder and every element is transformed by the XmlFilter from the lookup-table.

Documentation

transform :: XmlTree -> XmlArrow

filter for transforming the document.

1.parameter dtdPart : the DTD subset (Node DOCTYPE) of the XmlTree

2.parameter doc : the document subset of the XmlTree

returns : a list of errors