Data.Tree.NavigatableTree.Class

Description

Interface definition for navigatable trees. Navigatable trees need to have operations to move up, down, left and right. With these elementary operations, most of the XPath axises can be defined.

Documentation

class NavigatableTree t where

The interface for navigatable trees

Methods

mvUp :: t a -> Maybe (t a)

move one step towards the root

mvDown :: t a -> Maybe (t a)

descend one step to the leftmost child

mvLeft :: t a -> Maybe (t a)

move to the left neighbour

mvRight :: t a -> Maybe (t a)

move to the right neighbour

class NavigatableTreeToTree nt t | t -> nt, nt -> t where

Conversion between trees and navigatable trees,

There is only a single navigatable tree implementation for a given tree allowed (see the functional dependencies)

Methods

fromTree :: t a -> nt a

construct a navigatable tree

toTree :: nt a -> t a

remove navigation

class NavigatableTreeModify nt t | t -> nt, nt -> t where

Edit operation on navigatable trees

There is only a single navigatable tree implementation for a given tree allowed (see the functional dependencies)

Methods

addTreeLeft :: t a -> nt a -> Maybe (nt a)

add an ordinary tree in front of the given navigatable tree

addTreeRight :: t a -> nt a -> Maybe (nt a)

add an ordinary tree behind of the given navigatable tree

dropTreeLeft :: nt a -> Maybe (nt a)

drop the direct left sibling tree of the given navigatable tree

dropTreeRight :: nt a -> Maybe (nt a)

drop the direct right sibling tree of the given navigatable tree

substThisTree :: t a -> nt a -> nt a

change the tree but remain the navigation