Text.XML.HXT.Parser.XmlDTDParser

Description

Parsec parser for DTD declarations for ELEMENT, ATTLIST, ENTITY and NOTATION declarations

Documentation

parseXmlDTDdecl :: XmlTree -> XmlTrees

parse a tokenized DTD declaration represented by a DTD tree. The content is represented by the children containing text and parameter entity reference nodes. The parameter entity reference nodes contain their value in the children list, consisting of text and possibly again parameter entity reference nodes. This structure is build by the parameter entity substitution. Output is again a DTD declaration node, but this time completely parsed and ready for further DTD processing

parseXmlDTDdeclPart :: XmlTree -> XmlTrees

parseXmlDTDEntityValue :: XmlTree -> XmlTrees

elementDecl :: SParser XmlTrees

attlistDecl :: SParser XmlTrees

entityDecl :: SParser XmlTrees

notationDecl :: SParser XmlTrees