Text.XML.HXT.Parser.XmlParsec

Description

Xml Parsec parser with pure filter interface

Documentation

charData :: XParser s XmlTrees

charData' :: XParser s XmlTree

comment :: XParser s XmlTree

pI :: XParser s XmlTree

cDSect :: XParser s XmlTree

document :: XParser s XmlTree

document' :: XParser s XmlTrees

prolog :: XParser s XmlTrees

xMLDecl :: XParser s XmlTrees

xMLDecl' :: XParser s XmlTrees

versionInfo :: XParser s XmlTrees

misc :: XParser s XmlTree

doctypedecl :: XParser s XmlTrees

markupdecl :: XParser s XmlTrees

sDDecl :: XParser s XmlTrees

element :: XParser s XmlTree

content :: XParser s XmlTrees

contentWithTextDecl :: XParser s XmlTrees

textDecl :: XParser s XmlTrees

encodingDecl :: XParser s XmlTrees

xread :: String -> XmlTrees

the inverse function to xshow, (for XML content).

the string parameter is parsed with the XML content parser. result is the list of trees or in case of an error a single element list with the error message as node. No entity or character subtitution is done here, but the XML parser can do this for the predefined XML or the char references for performance reasons

see also: parseXmlContent

xreadDoc :: String -> XmlTrees

parseXmlContent :: XmlTree -> XmlTrees

the filter version of xread

parseXmlDocEncodingSpec :: XmlTree -> XmlTrees

parseXmlDocument :: String -> String -> XmlTrees

parseXmlDTDPart :: String -> XmlTree -> XmlTrees

Parser for parts of a DTD

parseXmlEncodingSpec :: SimpleXParser XmlTree -> XmlTree -> XmlTrees

try to parse a xml encoding spec.

1.parameter encParse : the parser for the encoding decl

2.parameter root : a document root

returns : the same tree, but with an additional attribute "encoding" in the root node in case of a valid encoding spec else the unchanged tree

parseXmlEntityEncodingSpec :: XmlTree -> XmlTrees

parseXmlEntityValueAsAttrValue :: String -> XmlTree -> XmlTrees

Parser for entity substitution within attribute values

parseXmlEntityValueAsContent :: String -> XmlTree -> XmlTrees

Parser for general entites

parseXmlPart :: SimpleXParser XmlTrees -> String -> String -> XmlTree -> XmlTrees

general parser for parsing arbitray parts of a XML document

parseXmlText :: SimpleXParser XmlTrees -> XPState () -> String -> XmlTree -> XmlTrees

a more general version of parseXmlContent. The parser to be used and the context are extra parameter

parseNMToken :: String -> XmlTree -> XmlTrees

Parser for NMTOKENs

parseName :: String -> XmlTree -> XmlTrees

Parser for XML names

removeEncodingSpec :: XmlTree -> XmlTrees